Boot does one browser round-trip for all preferences instead of four (each was a component mount that could force a rerun) — faster cold starts and no restore race. Balance keeps its cents. Overflow menu tightened: 32px rows, no stacked gaps, hairline dividers between View/Actions/Help

+++ src/edge_analysis/ui/tabs.py
@@ -598,7 +598,7 @@ def _perf_settings(g: pd.DataFrame):
             if "c" in _saved:
                 st.session_state["ea_m_cap"] = int(_saved["c"])
             if "b" in _saved:
-                st.session_state["ea_m_bal"] = int(_saved["b"])
+                st.session_state["ea_m_bal"] = float(_saved["b"])
                 st.session_state.pop("ea_m_bal_auto", None)  # it's yours now, not derived
         except (KeyError, TypeError, ValueError):
             pass
